refactor: 隔离爬虫实现细节
本次提交将爬虫专属模块从共享位置迁移至 `crawler` 包内部，以提升模块化程度和关注点分离。

- 将 `shared/database.py` 移至 `crawler/database.py`
- 将 `shared/video.py` 移至 `crawler/video.py`
- 删除`shared/`
- 修正导入路径
